finished imports for corona animation

+++ Animation_Service/WebApplication1/Models/CountryPopulation.cs
@@ -13,7 +13,7 @@ namespace WebApplication1.Models
         public int NetChange { get; set; }
         public int Density { get; set; }
         public int LandArea { get; set; }
-        public int Migrants { get; set; }
+        public double Migrants { get; set; }
         public double FertilityRate { get; set; }
         public int MedianAge { get; set; }
         public double UrbanPopulation { get; set; }
